+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-x.x.x Release notes (yyyy-MM-dd)
+2.5.1 Release notes (2017-04-05)
 =============================================================
 
 ### API Breaking Changes
@@ -7,8 +7,7 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 
-* Add a `{RLM}SyncUser.isAdmin` property indicating whether a user is a Realm
-  Object Server administrator.
+* None.
 
 ### Bugfixes